Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/svg/animations/slider-switch-expected.txt


FAIL Check correct event bases for onclick assert_equals: expected "block" but got "none"
